rustree::core::config

Struct RustreeLibConfig

pub struct RustreeLibConfig {

    pub root_display_name: String,
    pub max_depth: Option<usize>,
    pub show_hidden: bool,
    pub report_sizes: bool,
    pub report_permissions: bool,
    pub report_mtime: bool,
    pub calculate_line_count: bool,
    pub calculate_word_count: bool,
    pub apply_function: Option<BuiltInFunction>,
    pub sort_by: Option<SortKey>,
    pub reverse_sort: bool,
    pub list_directories_only: bool,
    pub root_node_size: Option<u64>,
    pub root_is_directory: bool,

}

Configuration for the rustree library.

This struct holds all the options that control how rustree processes and displays directory trees.

Fields
root_display_name: String

The name to display for the root of the scanned directory. Typically the directory name itself or โ.โ for the current directory.

max_depth: Option<usize>

The maximum depth to traverse into the directory structure. None means no limit. Some(0) would effectively show only the root (if walker adapted). Some(1) shows root and its direct children.

show_hidden: bool

If true, hidden files and directories (those starting with a .) will be included.

report_sizes: bool

If true, report the size of files.

report_permissions: bool

If true, report file permissions (currently not implemented in output).

report_mtime: bool

If true, report the last modification time of files and directories.

calculate_line_count: bool

If true, calculate and report the number of lines for files.

calculate_word_count: bool

If true, calculate and report the number of words for files.

apply_function: Option<BuiltInFunction>

Specifies a built-in function to apply to the content of each file. The result of this function can be displayed.

sort_by: Option<SortKey>

The key by which to sort directory entries. None means default (usually OS-dependent or DFS order).

reverse_sort: bool

If true, reverses the sort order specified by sort_by.

list_directories_only: bool

If true, only directories will be listed.

root_node_size: Option<u64>

Optional size of the root node itself, used by formatters if report_sizes is true.

root_is_directory: bool

Indicates if the root path itself is a directory.