rustree::core::formatter::base

Trait TreeFormatterCompat

pub trait TreeFormatterCompat: TreeFormatter {
    // Provided method
    fn format_compat(
        &self,
        nodes: &[NodeInfo],
        config: &RustreeLibConfig,
    ) -> Result<String, RustreeError> { ... }
}

Extension trait that provides backward compatibility with the old config-based API

Provided Methods

fn format_compat( &self, nodes: &[NodeInfo], config: &RustreeLibConfig, ) -> Result<String, RustreeError>

Formats the given nodes using the full config (backward compatibility)