rustree

Function create_context_from_options

pub fn create_context_from_options(
    listing: ListingOptions,
    filtering: FilteringOptions,
    metadata: MetadataOptions,
    input_source: InputSourceOptions,
    misc: MiscOptions,
    html: HtmlOptions,
    sorting: Option<SortingOptions>,
) -> OwnedProcessingContext

Create a processing context from individual option structs.

This function provides a convenient way to create processing contexts when you already have individual option structs, avoiding the need to use the builder pattern for simple cases.

Arguments

listing - Directory listing options

filtering - File filtering options

metadata - Metadata collection options

input_source - Input source options for formatting

misc - Miscellaneous options

html - HTML-specific options

sorting - Optional sorting options

Returns

A configured OwnedProcessingContext

Examples

use rustree::{create_context_from_options, ListingOptions, FilteringOptions};
use rustree::{MetadataOptions, InputSourceOptions, MiscOptions, HtmlOptions, SortingOptions};

let listing = ListingOptions { max_depth: Some(2), ..Default::default() };
let filtering = FilteringOptions::default();
let metadata = MetadataOptions { show_size_bytes: true, ..Default::default() };
let input_source = InputSourceOptions { root_display_name: "test".to_string(), ..Default::default() };
let misc = MiscOptions::default();
let html = HtmlOptions::default();
let sorting = Some(SortingOptions::default());

let context = create_context_from_options(
    listing, filtering, metadata, input_source, misc, html, sorting
);